[GRIFFIN-240] trigger returns triggerkey
This PR requires updating tests and postman json yet, but I need it to show the proposed API before moving further. Triggering and searching by triggerKey - both works.

Please look at the changes in `api-guide.md` for the proposed API additions.

This is rework of #489 (in asynchronous way, without timeout)

Author: Rodion Gorkovenko <aborgatin@griddynamics.com>
Author: Rodion Gorkovenko <rodiongork@github.com>

Closes #494 from RodionGork/GRIFFIN-240-trigger-returns-triggerkey.