[GRIFFIN-304] Eliminate older contexts
**What changes were proposed in this pull request?**

As SparkSession is a direct replacement for SparkContext, SQLContext and HiveContext, there is no need to pass/ instantiate them. If any of the oder contexts are needed, they can be derived from SparkSession.

This issue aims to eliminate dependency on older Contexts in favour of SparkSession.

**Does this PR introduce any user-facing change?**
No

**How was this patch tested?**
Griffin test suite.

Author: chitralverma <chitralverma@gmail.com>

Closes #557 from chitralverma/eliminate-older-contexts.