[GRIFFIN-317] Define guidelines for Griffin Project Improvement Proposals (GPIP)
**What changes were proposed in this pull request?**

Taking inspiration from Apache Spark, this ticket aims to define guidelines for Griffin Project Improvement Proposals (GPIP).

The purpose of a GPIP is to inform and involve the user community in major improvements to the Apache Griffin codebase throughout the development process, to increase the likelihood that user needs are met.

A GPIP aims to discuss the design and implementation of major features and changes in a collaborative manner. These major features must not be small/ incremental/ wide-scoped, as these features can be resolved by normal Jira process.

**Does this PR introduce any user-facing change?**
No

**How was this patch tested?**
Not Applicable

Author: chitralverma <chitralverma@gmail.com>

Closes #563 from chitralverma/griffin-pip-template.