optimize get metric maps in 'MetricWriteStep'
**Why/What changes？**
In 'MetricWriteStep.getMetricMaps()' the dataframe was transformed to json rdd, and then collect, and then transformed to Seq[Map].
It's not elegant and hard to understand. More optimized way is to collect it first, and then transform it to Seq[Map] directly.

We have test it with our DQ cases. It works well.

Author: yuxiaoyu <yuxiaoyu@bytedance.com>

Closes #566 from XiaoyuBD/optimizeMetricWriteGetMaps.