[GRIFFIN-316] Fix job exception handling
**What changes were proposed in this pull request?**

Currently we are using Try instance to represent the results of a DQ job, whether succeeded or failed. But as we are only wrapping the Boolean result by applying "Try" at the most outside level, the underlying failure would not be able to caught and it would always return "Success" even if exception got.

This is to modify all the underlying execute/doExecute methods of a DQ job, by handling exception with "Try" instances so that it could be passed properly to users when things get wrong.

**Does this PR introduce any user-facing change?**
No.

**How was this patch tested?**
Griffin test suite.

Author: Yu <yu.liu003@gmail.com>

Closes #562 from PnPie/exception_catch.