Support http remote conf
In our production practice, many Griffin jobs run on yarn in cluster mode. We upload  different conf files to the http file server and we also provide services that generate specific configurations based on different HTTP URLs.
So we supports setting HTTP URLs as conf in submitting Griffin jobs in this PR. There is no effect on JSON or File conf mode. And it works well in our production environment for a long time.

Author: yuxiaoyu <yuxiaoyu@bytedance.com>

Closes #587 from XiaoyuBD/support_http_url_conf.